# backend/src/api/routes/auth.py
"""Authentication routes for LexCore + LexRadar API."""

from datetime import datetime, timedelta
from typing import Optional

from fastapi import APIRouter, HTTPException, status, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pydantic import BaseModel, Field

from src.config.settings import settings

router = APIRouter()
security = HTTPBearer()


class TokenRequest(BaseModel):
    api_key: str = Field(..., min_length=32, max_length=128)


class TokenResponse(BaseModel):
    access_token: str
    refresh_token: str
    token_type: str = "Bearer"
    expires_in: int = 3600


class RefreshTokenRequest(BaseModel):
    refresh_token: str = Field(..., min_length=32)


@router.post("/token", response_model=TokenResponse)
async def get_token(request: TokenRequest) -> TokenResponse:
    """Exchange API key for JWT access token.
    
    Validates the API key against the database and returns
    a JWT access token and refresh token pair.
    """
    # TODO: Implement API key validation and JWT generation
    # 1. Hash API key and look up in database
    # 2. Verify key is active and not expired
    # 3. Check tenant tier for rate limits
    # 4. Generate JWT with tenant_id, role claims
    # 5. Generate refresh token
    # 6. Log token issuance to audit log
    
    if not request.api_key.startswith("lex_"):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid API key format",
        )
    
    return TokenResponse(
        access_token="jwt_token_placeholder",
        refresh_token="refresh_token_placeholder",
    )


@router.post("/refresh", response_model=TokenResponse)
async def refresh_token(request: RefreshTokenRequest) -> TokenResponse:
    """Refresh JWT access token using refresh token.
    
    Validates the refresh token and issues a new access token pair.
    Implements token rotation — old refresh token is invalidated.
    """
    # TODO: Implement refresh token validation and rotation
    # 1. Verify refresh token signature and expiry
    # 2. Check token is not revoked (Redis blacklist)
    # 3. Look up tenant_id from token claims
    # 4. Generate new JWT pair
    # 5. Blacklist old refresh token
    
    return TokenResponse(
        access_token="new_jwt_token_placeholder",
        refresh_token="new_refresh_token_placeholder",
    )


async def get_current_tenant(
    credentials: HTTPAuthorizationCredentials = Depends(security),
) -> dict:
    """Dependency to extract and validate JWT token.
    
    Returns tenant context for use in route handlers.
    Sets tenant_id in database session for RLS enforcement.
    """
    # TODO: Implement JWT validation
    # 1. Verify JWT signature with secret key
    # 2. Check token expiry
    # 3. Extract tenant_id from claims
    # 4. Verify tenant is active
    # 5. Set database session variable for RLS
    # 6. Log access for audit trail
    
    token = credentials.credentials
    
    if token == "jwt_token_placeholder":
        return {
            "id": "00000000-0000-0000-0000-000000000000",
            "tier": "ENTERPRISE",
            "name": "Test Tenant",
        }
    
    raise HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Invalid or expired token",
    )