# backend/src/api/routes/lexcore.py
"""LexCore legal intelligence routes."""

from typing import Optional
from uuid import UUID

from fastapi import APIRouter, Query, HTTPException, status, Depends

from src.api.models import MonitorRuleCreate, MonitorRule
from src.api.schemas import DocumentResponse
from src.api.middleware.jwt_auth import get_current_tenant

router = APIRouter()


@router.get("/documents")
async def list_documents(
    jurisdiction: Optional[str] = Query(None),
    body_of_law: Optional[str] = Query(None),
    limit: int = Query(50, ge=1, le=1000),
    offset: int = Query(0, ge=0),
    tenant: dict = Depends(get_current_tenant),
):
    """List legal documents with optional filtering."""
    # TODO: Implement document listing from database
    return {"documents": [], "total": 0, "limit": limit, "offset": offset}


@router.get("/documents/{id}", response_model=DocumentResponse)
async def get_document_by_id(
    id: UUID,
    tenant: dict = Depends(get_current_tenant),
) -> DocumentResponse:
    """Get a specific legal document by ID."""
    # TODO: Implement document retrieval
    return DocumentResponse(document={}, chunks=[], citations=[])


@router.get("/chunks")
async def list_chunks(
    document_id: Optional[UUID] = Query(None),
    limit: int = Query(50, ge=1, le=1000),
    tenant: dict = Depends(get_current_tenant),
):
    """List document chunks with optional document filter."""
    # TODO: Implement chunk listing
    return {"chunks": [], "total": 0}


@router.post("/monitor-rules", response_model=MonitorRule, status_code=status.HTTP_201_CREATED)
async def create_monitor_rule(
    rule: MonitorRuleCreate,
    tenant: dict = Depends(get_current_tenant),
) -> MonitorRule:
    """Create a new monitor rule for legislative changes."""
    # TODO: Implement monitor rule creation
    return MonitorRule(
        id=UUID("00000000-0000-0000-0000-000000000000"),
        tenant_id=tenant["id"],
        **rule.model_dump(),
        last_checked=None,
        created_at=__import__("datetime").datetime.utcnow(),
        updated_at=__import__("datetime").datetime.utcnow(),
    )


@router.delete("/monitor-rules/{id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_monitor_rule(
    id: UUID,
    tenant: dict = Depends(get_current_tenant),
):
    """Delete a monitor rule."""
    # TODO: Implement monitor rule deletion
    pass