# backend/src/api/routes/lexradar.py
"""LexRadar IP detection routes."""

from typing import Optional
from uuid import UUID

from fastapi import APIRouter, Query, HTTPException, status, Depends

from src.api.models import InventionCreate, Invention
from src.api.schemas import (
    DisclosureRequest,
    DisclosureResponse,
    FilingBundleResponse,
    LedgerProofResponse,
)
from src.api.middleware.jwt_auth import get_current_tenant

router = APIRouter()


@router.get("/inventions")
async def list_inventions(
    status: Optional[str] = Query(None, pattern=r"^(DETECTED|SCORING|DISCLOSED|FILED)$"),
    limit: int = Query(50, ge=1, le=1000),
    tenant: dict = Depends(get_current_tenant),
):
    """List inventions with optional status filter."""
    # TODO: Implement invention listing
    return {"inventions": [], "total": 0}


@router.post("/inventions", response_model=Invention, status_code=status.HTTP_201_CREATED)
async def create_invention(
    invention: InventionCreate,
    tenant: dict = Depends(get_current_tenant),
) -> Invention:
    """Create a new invention record."""
    # TODO: Implement invention creation
    return Invention(
        id=UUID("00000000-0000-0000-0000-000000000000"),
        tenant_id=tenant["id"],
        **invention.model_dump(),
        detected_at=__import__("datetime").datetime.utcnow(),
        disclosed_at=None,
        filed_at=None,
        created_at=__import__("datetime").datetime.utcnow(),
        updated_at=__import__("datetime").datetime.utcnow(),
    )


@router.post("/inventions/{id}/prior-art")
async def search_prior_art(
    id: UUID,
    sources: list[str] = Query(...),
    max_results: int = Query(50, ge=1, le=500),
    tenant: dict = Depends(get_current_tenant),
):
    """Search prior art for an invention across 7 sources."""
    # TODO: Implement parallel prior art search
    return {"prior_art": [], "total_found": 0}


@router.post("/inventions/{id}/disclosure", response_model=DisclosureResponse, status_code=status.HTTP_201_CREATED)
async def generate_disclosure(
    id: UUID,
    request: DisclosureRequest,
    tenant: dict = Depends(get_current_tenant),
) -> DisclosureResponse:
    """Generate a disclosure draft for an invention."""
    # TODO: Implement disclosure generation with grounding check
    return DisclosureResponse(
        id=UUID("00000000-0000-0000-0000-000000000000"),
        invention_id=id,
        disclosure_type=request.disclosure_type,
        sections={},
        claim_themes=[],
        grounding_score=0.0,
        status="DRAFT",
    )


@router.post("/inventions/{id}/filing-bundle", response_model=FilingBundleResponse, status_code=status.HTTP_201_CREATED)
async def package_filing_bundle(
    id: UUID,
    tenant: dict = Depends(get_current_tenant),
) -> FilingBundleResponse:
    """Package a filing bundle with 9 documents."""
    # TODO: Implement bundle packaging
    return FilingBundleResponse(
        id=UUID("00000000-0000-0000-0000-000000000000"),
        disclosure_draft_id=UUID("00000000-0000-0000-0000-000000000000"),
        documents=[],
        bundle_path="",
        bundle_format="PDF",
        status="PACKAGED",
    )


@router.get("/ledger/{id}", response_model=LedgerProofResponse)
async def get_ledger_proof(
    id: UUID,
    tenant: dict = Depends(get_current_tenant),
) -> LedgerProofResponse:
    """Get ledger proof for a filing bundle."""
    # TODO: Implement ledger proof retrieval
    return LedgerProofResponse(
        id=id,
        document_hash="",
        bundle_hash="",
        polygon_tx_hash="",
        polygon_block_number="",
        anchored_at=__import__("datetime").datetime.utcnow(),
    )